COMET: Fix GCC Warnings.
git-svn-id: https://svn.scummvm.org:4444/svn/comet@197 e395a8f9-5cf0-422e-971d-9a083d8f24e8

+++ engines/comet/animationmgr.cpp
@@ -67,7 +67,7 @@ AnimationResource *AnimationManager::loadAnimationResource(const char *pakFilena
 }
 
 void AnimationManager::purgeUnusedAnimationSlots() {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
 		if (_animationSlots[i].anim && _animationSlots[i].animationType == 0 && !_vm->isAnimationSlotUsed(i)) {
 			_vm->clearAnimationSlotByIndex(i);
 			delete _animationSlots[i].anim;
@@ -77,8 +77,8 @@ void AnimationManager::purgeUnusedAnimationSlots() {
 }
 
 void AnimationManager::purgeAnimationSlots() {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
-		if (_animationSlots[i].anim && _animationSlots[i].animationType == 0 && _vm->_actors[0].animationSlot != i) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
+		if (_animationSlots[i].anim && _animationSlots[i].animationType == 0 && _vm->_actors[0].animationSlot != (int) i) {
 			_vm->clearAnimationSlotByIndex(i);
 			delete _animationSlots[i].anim;
 			_animationSlots[i].anim = NULL;
@@ -110,7 +110,7 @@ int AnimationManager::getAnimationResource(int16 animationType, int16 fileIndex)
 }
 
 void AnimationManager::refreshAnimationSlots() {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
 		if (_animationSlots[i].anim && _animationSlots[i].animationType == 0) {
 			delete _animationSlots[i].anim;
 			_animationSlots[i].anim = NULL;
@@ -120,7 +120,7 @@ void AnimationManager::refreshAnimationSlots() {
 }
 
 void AnimationManager::restoreAnimationSlots() {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
 		if (_animationSlots[i].fileIndex != -1) {
 			if (_animationSlots[i].animationType == 0) {
 				delete _animationSlots[i].anim;
@@ -135,7 +135,7 @@ void AnimationManager::restoreAnimationSlots() {
 }
 
 int AnimationManager::findAnimationSlot(int16 animationType, int16 fileIndex) {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
 		if (_animationSlots[i].animationType == animationType && _animationSlots[i].fileIndex == fileIndex) {
 			return i;
 		}
@@ -144,7 +144,7 @@ int AnimationManager::findAnimationSlot(int16 animationType, int16 fileIndex) {
 }
 
 int AnimationManager::findFreeAnimationSlot() {
-	for (int i = 0; i < kAnimationSlotCount; i++) {
+	for (uint i = 0; i < kAnimationSlotCount; i++) {
 		if (_animationSlots[i].anim == NULL) {
 			return i;
 		}
